refactor: move manager aggregates to another file
