+++ drift/lib/src/runtime/manager/manager.dart
@@ -13,6 +13,7 @@ part 'join_builder.dart';
 part 'ordering.dart';
 part 'references.dart';
 part 'annotate.dart';
+part 'aggregates.dart';
 
 /// Defines a class that holds the state for a [BaseTableManager]
 ///
